implementation of "DomaineDAO" and "ReferentielDAO" and test for "DomaineDAO" replace "int" in "long" for id in interface of "domaine" and "referentiel"

+++ sadoc-parent/sadoc-dao-parent/sadoc-dao-commun/src/main/java/fr/univartois/ili/sadoc/dao/services/IDomaineDAO.java
@@ -18,7 +18,7 @@ public interface IDomaineDAO {
 	 * @return Domaine
 	 * return the corresponding domaine
 	 */
-	Domaine findDomaineById(int id);
+	Domaine findDomaineById(long id);
 
 	/**
 	 * Insert the domaine in DB
